requirements.txt
flask==2.3.3
feedparser==6.0.10
requests==2.31.0
anthropic==0.21.0
beautifulsoup4==4.12.2
nltk==3.8.1
python-dotenv==1.0.0
